pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.1.1",
    "python-telegram-bot==20.8",
    "supabase>=2.18.1",
    "telegram>=0.0.1",
]
